requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
librosa>=0.10
soundfile
numpy<2.0
scikit-learn
matplotlib
tqdm
tensorboard
omegaconf
pandas
audiomentations
pydub
pyyaml